fix: analyzer errors — int/double types, mounted check, unnecessary !
- budgetNearPct: int → double
- weekendRatioPct: int → double
- Removed unnecessary ! on non-nullable String deadline
- Fixed use_build_context_synchronously in pin toggle

+++ lib/models/recommendation_prefs.dart
@@ -13,7 +13,7 @@ class RecommendationPrefs {
   double emergencyMonths;
   double idleCashMin;
   int noBudgetMinSpend;
-  int budgetNearPct;
+  double budgetNearPct;
   double topCatMinPct;
 
   // --- New types ---
@@ -21,7 +21,7 @@ class RecommendationPrefs {
   double spikePct;
   int recurringMonths;
   double singleCatDominancePct;
-  int weekendRatioPct;
+  double weekendRatioPct;
   double largeCashMin;
 
   RecommendationPrefs({
